pms/requirements.txt
Django==6.0.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.0
setuptools<81
